Update ResourceList tests
